refactor(buffers): remove multi-stage buffer topology and overflow mode
Drop the `WhenFull::Overflow` variant and the ability to chain multiple
buffer stages. Buffers now support exactly two modes—block and drop
newest—with a single stage (memory or disk).

This removes a significant amount of structural complexity that was
never promoted beyond `docs::hidden`:

- `BufferConfig` flattened from enum (`Single`/`Chained`) to newtype
- `BufferSender`/`BufferReceiver` no longer carry recursive overflow
  pointers, eliminating the `async-recursion` dependency
- `TopologyBuilder::build()` simplified from reverse-iteration
  multi-stage wiring to single-stage construction
- `stages()` replaced with `buffer_type()` across all call sites

+++ src/topology/test/backpressure.rs
@@ -107,7 +107,7 @@ async fn buffer_drop_fan_out() {
         vec!["in".to_string()],
         backpressure_sink(events_to_sink / 2),
     );
-    sink_outer.buffer = BufferConfig::Single(BufferType::Memory {
+    sink_outer.buffer = BufferConfig(BufferType::Memory {
         size: MemoryBufferSize::MaxEvents(MEMORY_BUFFER_DEFAULT_MAX_EVENTS),
         when_full: WhenFull::DropNewest,
     });
